+++ tools/src/main/java/org/esa/cci/sst/regavg/RegionalAveraging.java
@@ -65,7 +65,7 @@ public class RegionalAveraging {
                 Date date2 = calendar.getTime();
                 List<OutputTimeStep> intermediateResults = computeOutputTimeSteps(fileStore, climatology, outputType, sstDepth, date1, date2, TemporalResolution.monthly, regionMaskList);
                 result = aggregateMultiMonthsResults(intermediateResults, regionMaskList);
-            } else /*if (temporalResolution == TemporalResolution.anual)*/ {
+            } else /*if (temporalResolution == TemporalResolution.annual)*/ {
                 calendar.add(Calendar.YEAR, 1);
                 Date date2 = calendar.getTime();
                 List<OutputTimeStep> intermediateResults = computeOutputTimeSteps(fileStore, climatology, outputType, sstDepth, date1, date2, TemporalResolution.monthly, regionMaskList);
@@ -76,14 +76,19 @@ public class RegionalAveraging {
         return results;
     }
 
-    private static List<Cell> aggregateMultiMonthsResults(List<OutputTimeStep> monthlyResults, RegionMaskList regionMaskList) {
-        // todo - the following code is only valid for simple mean (SST, SST anomaly), but nor for the uncertaincies. Aggregate these according to Nick's equations
+    private static List<Cell> aggregateMultiMonthsResults(List<OutputTimeStep> monthlyResults,
+                                                          RegionMaskList regionMaskList) {
+        // todo - the following code is only valid for simple mean (SST, SST anomaly), but not for the uncertainties.
+        // Aggregate these according to Nick's equations:
+        // Variable A, Variable B, Variable C, Variable E, Variable F, Variable G  – according to Equation 1.1
+        // Variable D  – according to Equation 1.2
+        //
         ArrayList<Cell> combinedCells = new ArrayList<Cell>();
         for (int i = 0; i < regionMaskList.size(); i++) {
             Cell combinedCell = new Cell();
             for (OutputTimeStep monthlyResult : monthlyResults) {
                 Cell cell = monthlyResult.regionalAverages.get(i);
-                combinedCell.accumulate(cell);
+                combinedCell.accumulateCellAverage(cell, 1.0);
             }
             combinedCells.add(combinedCell);
         }
@@ -100,39 +105,37 @@ public class RegionalAveraging {
 
         RegionMask combinedRegionMask = RegionMask.combine(regionMaskList);
 
-        CellGrid combined5DGrid = computeCombinedGrid(fileStore, climatology, outputType, sstDepth, date1, date2, combinedRegionMask);
+        CellGrid combined5DGrid = aggregateTimeRangeTo5DegGrid(fileStore, climatology, outputType, sstDepth, date1, date2, combinedRegionMask);
 
         List<Cell> regionalCells = new ArrayList<Cell>();
         for (RegionMask regionMask : regionMaskList) {
 
             // If we have multiple masks, split the aggregation results we've got to each region grid.
-            CellGrid regional5DegGrid = getRegionalGrid(combined5DGrid, climatology.getSeaCoverage5DegGrid(), regionMask);
+            CellGrid regional5DegGrid = getRegionalGrid(combined5DGrid, regionMask);
 
             // Check if region is Globe or Hemisphere (actually: check if we have fully covered
             // 90 deg grid boxes), if so apply special averaging for all 90 deg grid boxes.
             if (regionMask.getCoverage() == RegionMask.Coverage.Globe
                     || regionMask.getCoverage() == RegionMask.Coverage.N_Hemisphere
                     || regionMask.getCoverage() == RegionMask.Coverage.S_Hemisphere) {
-                CellGrid regional90DegGrid = averageTo90DegGrid(regional5DegGrid, climatology.getSeaCoverage90DegGrid(), combinedRegionMask);
-                regionalCells.add(regional90DegGrid.combine());
+                CellGrid regional90DegGrid = aggregate5DegTo90DegGrid(regional5DegGrid, combinedRegionMask);
+                regionalCells.add(regional90DegGrid.aggregate(climatology.getSeaCoverage90DegGrid()));
             } else {
-                regionalCells.add(regional5DegGrid.combine());
+                regionalCells.add(regional5DegGrid.aggregate(climatology.getSeaCoverage5DegGrid()));
             }
         }
 
         return regionalCells;
     }
 
-    private static CellGrid getRegionalGrid(CellGrid combined5DegGrid, Grid waterCoverage5DegGrid, RegionMask regionMask) {
+    private static CellGrid getRegionalGrid(CellGrid combined5DegGrid, RegionMask regionMask) {
         CellGrid regional5DGrid = new CellGrid(combined5DegGrid.getGridDef());
         for (int cellY = 0; cellY < combined5DegGrid.getGridDef().getHeight(); cellY++) {
             for (int cellX = 0; cellX < combined5DegGrid.getGridDef().getWidth(); cellX++) {
                 if (regionMask.getSampleBoolean(cellX, cellY)) {
                     Cell cell = combined5DegGrid.getCell(cellX, cellY);
-                    if (cell != null && cell.getAccuCount() > 0) {
-                        Cell cellClone = cell.clone();  // :-)
-                        cellClone.setWeight(waterCoverage5DegGrid.getSampleDouble(cellX, cellY));
-                        regional5DGrid.setCell(cellX, cellY, cellClone);
+                    if (cell != null && cell.getSampleCount() > 0) {
+                        regional5DGrid.setCell(cellX, cellY, cell.clone());
                     }
                 }
             }
@@ -140,18 +143,17 @@ public class RegionalAveraging {
         return regional5DGrid;
     }
 
-    private static CellGrid averageTo90DegGrid(CellGrid regional5DegGrid, Grid waterCoverage90DegGrid, RegionMask combinedRegionMask) {
+    private static CellGrid aggregate5DegTo90DegGrid(CellGrid regional5DegGrid, RegionMask combinedRegionMask) {
         CellGrid regional90DegGrid = new CellGrid(GLOBAL_90_DEG_GRID_DEF);
         for (int cellY = 0; cellY < regional5DegGrid.getGridDef().getHeight(); cellY++) {
             for (int cellX = 0; cellX < regional5DegGrid.getGridDef().getWidth(); cellX++) {
                 if (combinedRegionMask.getSampleBoolean(cellX, cellY)) {
                     Cell regional5DegCell = regional5DegGrid.getCell(cellX, cellY);
-                    if (regional5DegCell != null && regional5DegCell.getAccuCount() > 0) {
+                    if (regional5DegCell != null && regional5DegCell.getSampleCount() > 0) {
                         int cellX90D = (cellX * regional90DegGrid.getGridDef().getWidth()) / regional5DegGrid.getGridDef().getWidth();
                         int cellY90D = (cellY * regional90DegGrid.getGridDef().getHeight()) / regional5DegGrid.getGridDef().getHeight();
                         Cell regional90DCell = regional90DegGrid.getCellSafe(cellX90D, cellY90D);
-                        regional90DCell.setWeight(waterCoverage90DegGrid.getSampleDouble(cellX90D, cellY90D));
-                        regional90DCell.accumulate(regional5DegCell);
+                        regional90DCell.accumulateCellSamples(regional5DegCell);
                     }
                 }
             }
@@ -159,35 +161,35 @@ public class RegionalAveraging {
         return regional90DegGrid;
     }
 
-    private static CellGrid computeCombinedGrid(FileStore fileStore,
-                                                Climatology climatology,
-                                                OutputType outputType, SstDepth sstDepth,
-                                                Date date1, Date date2,
-                                                RegionMask combinedRegionMask) throws IOException {
+    private static CellGrid aggregateTimeRangeTo5DegGrid(FileStore fileStore,
+                                                         Climatology climatology,
+                                                         OutputType outputType, SstDepth sstDepth,
+                                                         Date date1, Date date2,
+                                                         RegionMask combinedRegionMask) throws IOException {
         List<File> files = fileStore.getFiles(date1, date2);
 
         DateFormat isoDateFormat = UTC.getIsoFormat();
         LOGGER.info(String.format("Computing output time step from %s to %s, %d file(s) found.",
                                   isoDateFormat.format(date1), isoDateFormat.format(date2), files.size()));
 
-        CellGrid combined5DGrid = new CellGrid(GLOBAL_5_DEG_GRID_DEF);
+        CellGrid combined5DegGrid = new CellGrid(GLOBAL_5_DEG_GRID_DEF);
 
         for (File file : files) {
             LOGGER.info(String.format("Processing input %s file '%s'", fileStore.getProductType(), file));
             long t0 = System.currentTimeMillis();
             NetcdfFile netcdfFile = NetcdfFile.open(file.getPath());
             try {
-                aggregateFile(fileStore.getProductType(), netcdfFile, climatology, outputType, sstDepth, combinedRegionMask, combined5DGrid);
+                aggregateFileTo5DegGrid(fileStore.getProductType(), netcdfFile, climatology, outputType, sstDepth, combinedRegionMask, combined5DegGrid);
             } finally {
                 netcdfFile.close();
             }
             LOGGER.fine(String.format("Processing input %s file took %d ms", fileStore.getProductType(), System.currentTimeMillis() - t0));
         }
 
-        return combined5DGrid;
+        return combined5DegGrid;
     }
 
-    private static void aggregateFile(ProductType productType, NetcdfFile netcdfFile, Climatology climatology, OutputType outputType, SstDepth sstDepth, RegionMask combinedRegionMask, CellGrid combined5DGrid) throws IOException {
+    private static void aggregateFileTo5DegGrid(ProductType productType, NetcdfFile netcdfFile, Climatology climatology, OutputType outputType, SstDepth sstDepth, RegionMask combinedRegionMask, CellGrid combined5DGrid) throws IOException {
 
         Date date = productType.getFileType().readDate(netcdfFile);
         int dayOfYear = UTC.getDayOfYear(date);
@@ -200,7 +202,7 @@ public class RegionalAveraging {
         Grid analysedSstGrid = climatology.getAnalysedSstGrid(dayOfYear);
         Grid seaCoverageGrid = climatology.getSeaCoverageGrid();
 
-        aggregate(sstGrid, analysedSstGrid, seaCoverageGrid, combinedRegionMask, combined5DGrid, outputType);
+        accumulateInputTo5DegGrid(sstGrid, analysedSstGrid, seaCoverageGrid, combinedRegionMask, combined5DGrid, outputType);
     }
 
     private static Grid[] readGrids(ProductType productType, NetcdfFile netcdfFile, SstDepth sstDepth) throws IOException {
@@ -215,7 +217,12 @@ public class RegionalAveraging {
         return grids;
     }
 
-    private static void aggregate(Grid sstGrid, Grid analysedSstGrid, Grid seaCoverageGrid, RegionMask combinedRegionMask, CellGrid combined5DGrid, OutputType outputType) {
+    private static void accumulateInputTo5DegGrid(Grid sstGrid,
+                                                  Grid analysedSstGrid,
+                                                  Grid seaCoverageGrid,
+                                                  RegionMask combinedRegionMask,
+                                                  CellGrid combined5DegGrid,
+                                                  OutputType outputType) {
         long t0 = System.currentTimeMillis();
         LOGGER.fine("Aggregating grid(s)...");
         GridDef sourceGridDef = sstGrid.getGridDef();
@@ -225,12 +232,8 @@ public class RegionalAveraging {
 
                     Rectangle2D lonLatRectangle = combinedRegionMask.getGridDef().getLonLatRectangle(cellX, cellY);
                     Rectangle sourceGridRectangle = sourceGridDef.getGridRectangle(lonLatRectangle);
-
-                    Cell combined5DCell = combined5DGrid.createCell();
-                    aggregateRect(sstGrid, analysedSstGrid, seaCoverageGrid, sourceGridRectangle, outputType, combined5DCell);
-                    if (combined5DCell.getAccuCount() > 0) {
-                        combined5DGrid.setCell(cellX, cellY, combined5DCell);
-                    }
+                    Cell combined5DegCell = combined5DegGrid.getCellSafe(cellX, cellY) ;
+                    accumulate5DegCell(sstGrid, analysedSstGrid, seaCoverageGrid, sourceGridRectangle, outputType, combined5DegCell);
 
                     // todo - for each uncertainty variables A-G: create a grid comprising max. 72 x 36 aggregation results.
                 }
@@ -239,7 +242,7 @@ public class RegionalAveraging {
         LOGGER.fine(String.format("Aggregating grid(s) took %d ms", (System.currentTimeMillis() - t0)));
     }
 
-    public static void aggregateRect(Grid grid, Grid refGrid, Grid weightGrid, Rectangle rect, OutputType outputType, Cell cell) {
+    public static void accumulate5DegCell(Grid grid, Grid refGrid, Grid weightGrid, Rectangle rect, OutputType outputType, Cell cell) {
         final int x0 = rect.x;
         final int y0 = rect.y;
         final int x1 = x0 + rect.width - 1;
@@ -250,9 +253,9 @@ public class RegionalAveraging {
                 final double weight = weightGrid.getSampleDouble(x, y);
                 if (weight != 0.0) {
                     if (anomaly) {
-                        cell.accumulate(grid.getSampleDouble(x, y) - refGrid.getSampleDouble(x, y), weight, 1L);
+                        cell.accumulateSample(grid.getSampleDouble(x, y) - refGrid.getSampleDouble(x, y), weight);
                     } else {
-                        cell.accumulate(grid.getSampleDouble(x, y), weight, 1L);
+                        cell.accumulateSample(grid.getSampleDouble(x, y), weight);
                     }
                 }
             }